feat: add data helpers for class and lesson plan linking
