+++ src/lib/lessonPlans.ts
@@ -50,15 +50,6 @@ export interface LessonPlanWithSteps {
   steps: LessonStep[];
 }
 
-export interface UpcomingLessonPlanSummary {
-  id: string;
-  classId: string;
-  classTitle: string;
-  lessonPlanId: string;
-  lessonTitle: string;
-  date: string | null;
-}
-
 async function requireUserId(client: Client, action: string): Promise<string> {
   const { data, error } = await client.auth.getSession();
 
@@ -387,74 +378,6 @@ export async function getMyPlans(
   }
 
   return Array.isArray(data) ? data.map(mapLessonPlan) : [];
-}
-
-export async function listUpcomingLessonPlans(
-  limit = 5,
-  client: Client = supabase,
-): Promise<UpcomingLessonPlanSummary[]> {
-  await requireUserId(client, "view upcoming lesson plans");
-
-  const today = new Date().toISOString().slice(0, 10);
-  const requestedLimit = Number.isFinite(limit) && limit > 0 ? Math.floor(limit) : 5;
-  const safeLimit = Math.min(Math.max(requestedLimit, 1), 20);
-
-  const { data, error } = await client
-    .from("class_lesson_plans")
-    .select(
-      `
-        id,
-        class_id,
-        lesson_plan_id,
-        lesson_plans!inner (
-          id,
-          title,
-          date
-        ),
-        classes!inner (
-          id,
-          title
-        )
-      `,
-    )
-    .gte("lesson_plans.date", today)
-    .order("date", { ascending: true, nullsFirst: false, foreignTable: "lesson_plans" })
-    .limit(safeLimit);
-
-  if (error) {
-    throw new LessonPlanDataError("Failed to load upcoming lesson plans.", { cause: error });
-  }
-
-  if (!Array.isArray(data)) {
-    return [];
-  }
-
-  return data.map(record => {
-    const lessonPlan = (record as Record<string, any>).lesson_plans ?? null;
-    const classRecord = (record as Record<string, any>).classes ?? null;
-
-    const lessonTitle =
-      lessonPlan && typeof lessonPlan.title === "string" && lessonPlan.title.trim().length > 0
-        ? lessonPlan.title.trim()
-        : "Untitled lesson";
-    const classTitle =
-      classRecord && typeof classRecord.title === "string" && classRecord.title.trim().length > 0
-        ? classRecord.title.trim()
-        : "Untitled class";
-
-    const rawDate = lessonPlan?.date;
-    const formattedDate =
-      typeof rawDate === "string" && rawDate.trim().length > 0 ? rawDate.trim() : null;
-
-    return {
-      id: String(record.id ?? ""),
-      classId: String(record.class_id ?? classRecord?.id ?? ""),
-      lessonPlanId: String(record.lesson_plan_id ?? lessonPlan?.id ?? ""),
-      classTitle,
-      lessonTitle,
-      date: formattedDate,
-    } satisfies UpcomingLessonPlanSummary;
-  });
 }
 
 export async function getPlanWithSteps(
